+++ lib/carrierwave/active_graph/uploader_converter.rb
@@ -1,4 +1,4 @@
-module CarrierWave::Neo4j
+module CarrierWave::ActiveGraph
   module TypeConverters
     class UploaderConverter
       class << self
@@ -23,7 +23,7 @@ module CarrierWave::Neo4j
 end
 
 def register_converter(mod)
-  mod::TypeConverters.send :include, CarrierWave::Neo4j::TypeConverters
+  mod::TypeConverters.send :include, CarrierWave::ActiveGraph::TypeConverters
 end
 
 major = ActiveGraph::VERSION.split('.').first.to_i
@@ -32,5 +32,5 @@ case major
   when 1..2 then fail('Unsupported version of Neo4j gem. Please update it.')
   when 3    then register_converter(Neo4j)
   when 4    then register_converter(Neo4j::Shared)
-  when 5..Float::INFINITY then ActiveGraph::Shared::TypeConverters.register_converter(CarrierWave::Neo4j::TypeConverters::UploaderConverter)
+  when 5..Float::INFINITY then ActiveGraph::Shared::TypeConverters.register_converter(CarrierWave::ActiveGraph::TypeConverters::UploaderConverter)
 end
